Chain sample (#152)
Added sample that shows how to connect two chain shapes.
b2Rot now has cosine first.
Fixed polygon versus smooth segment collision for concave portion.
Shapes are now connected to bodies using a doubly linked list for fast
removal.
API structures now using int32_t instead of int for easier binding.

fixes #117
fixes #105

+++ samples/sample_benchmark.cpp
@@ -1269,7 +1269,7 @@ public:
 		b2ShapeDef shapeDef = b2DefaultShapeDef();
 
 		const char* text = "I";
-		int n = strlen(text);
+		int n = (int)strlen(text);
 		float zoom = 1.0f;
 
 		float x = 0.0f;
